Updated shaders & Fixing CameraControls
Made changes to the shaders to prevent an Invisible error due to chars I could not see.
Started the process now of converting the Camera class over to Matrices. Currently not having much success in that area but the shaders are set now so that if and when the default fixed function pipeline kicks in due to internal error it will be blatantly obvious that it has done so.